move auth and aws helper
